Add sidebar mobile as drawer with tabs
