--- app/src/Modals/sidebarMobile.scss
@@ -1,2 +1,0 @@
-@import "../scss/_variables.scss";
-@import "./Modals.scss";
