+++ package.json
@@ -26,6 +26,7 @@
     "@gnosis.pm/conditional-tokens-market-makers": "^1.3.0",
     "@hot-loader/react-dom": "^16.10.2",
     "@material-ui/core": "^4.6.0",
+    "@material-ui/icons": "^4.5.1",
     "@truffle/contract": "^4.0.33",
     "@use-it/interval": "^0.1.3",
     "canonical-weth": "^1.4.0",
@@ -46,7 +47,7 @@
     "react-media": "^1.10.0",
     "react-redux": "^7.1.1",
     "react-router-dom": "^5.1.2",
-    "react-select": "^3.0.4",
+    "react-select": "^3.0.8",
     "redux": "^4.0.4",
     "truffle-hdwallet-provider": "^1.0.17",
     "web3": "1.2.1"
